Refactor SEG-Y sampling interval helpers
